requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.4
requests==2.31.0
python-dotenv==1.0.1
docx2txt==0.8
PyMuPDF==1.24.1
